[MOGG][Extensibility] Translate ScatterND Ops to new API (#47477)
Stacked PRs:
 * __->__#47477

--- --- ---

### [MOGG][Extensibility] Translate ScatterND Ops to new API

Towards GRA-740.

Note, while there is a shape function for this operator, it is not
needed anymore. This is because the output shape is always the same as
the input shape, some runtime checks belong in the kernel.

MOGG_ORIG_REV_ID: 8a7ed2bb6d8ba0f02dc6fe53e3638861b768c75a