[AsyncRT] Implement DeviceBuffer::ptr() (#47488)
This allows raw access to the device pointer for a given buffer. We
support this reluctantly, since existing users expect to be able to
capture a reference to a DeviceBuffer in a function that is later
compiled for the device.

GPU_ORIG_REV_ID: ea8be99a7deca1836158dfc5d19eb57d552d8598