[mojo-lang][CheckLifetimes] Rewrite dtor insertion (#47639)
This is a substantial rewrite of the destructor insertion pass of
checkliftimes.
Previously, this code was trying to do two things at once: walking the
IR to
understand where to insert destructor calls, and also doing all the
complicated
logic to do dtor insertion + copyinit elision.

This is a problem because the code would trigger these rewrites as the
operands to the operation are walked, and in the case of a copyinit,
dtor
elision could go and nuke out the current operation.

This patch splits this up so that the operand walking loop decides
/what/
should be destroyed and enqueues it into a new DestructorInserter
instance, and then that thing all the generated destructor calls are
emitted
at once when the full set is calculated. This makes the. code
incrementally
easier to understand (due to separation of concerns), makes it more
extensible to future improvements, and provides a natural place to order
the generation of destructor calls.

This patch increases the effectiveness of copy elision a bit and
slightly
changes the order of dtor emission when one operation is the last use of
multiple values, but is otherwise just a giant refactor.

GPU_TEST_ORIG_REV_ID: 46b91ac692b5ac75a0f80a78113519a32aa19983