[Kernels] KV cache fixes for tiny llama tests (#47733)
Couple of fixes to work with the TinyLlama test cases:
1. TinyLlama is a model with depth_dim=10. The rope kernel required that
the depth_dim be a multiple of the SIMD width, but this can be relaxed
to `kernel_width=math.gcd(target_width, depth_dim)` as long as
`kernel_width>=2`.
2. Several of the flash attention kernel exports required a 2D mask
tensor, but the actual kernel (as well as the exports used for llama3)
don't have this constraint. All the exports are consistent now.

With these, the model successfully compiles but then hits failures in
the python tests.

Resolves KERN-984.

NN_ORIG_REV_ID: 094f5635d2f770a6d6c75214dde2734b6aa08d70