[NFC]Revert "[GRA-1047] [NFC] Repro test (#47699)" (#47749)
This reverts commit 03d04f812297c28208168bac61aa1521ce6ee7c8.

This test did not quite indicate the problem we thought that it
indicated. We do need to better handle the failure case when we try to
use `bfloat16` values on ARM CPUs. We will handle that in another issue.

MAX_GRAPH_API_ORIG_REV_ID: b1ac8424a74315dfc3cf546218eb64224e5850bc