[mojo-stdlib] Parameterize `UnsafePointer` on a lifetime.
This adds a new lifetime parameter to `UnsafePointer`, which specifies
the lifetime of the dereferenced pointer.  It defaults to `AnyLifetime`.

This has two advantages:

1) This can be used by unsafe code that wants to be more specific about
lifetimes than using `AnyLifetime`. This is important as we continue to
    push lifetimes in the stack.

2) This allows the compiler to know that unqualified `UnsafePointer`'s
involve `AnyLifetime` which means that CheckLifetimes understand that
APIs using `UnsafePointer` may access any live value. This allows it to
extend the lifetime of local values, which defines away another class of
    reasons that people have to use `_ = foo` to extend lifetimes.

Similar treatment should be applied to the `Buffer`/`StringRef` and
other similar types that are wrappers around `UnsafePointer`

LINALG_ORIG_REV_ID: 9c4ec32a5c57fe1c157ae1ac752c71279ada476e