[Kernels] Add initial binders for KVCache construction (#47957)
This change will improve on the KVCacheCollection construction which is
currently accomplished via a subgraph (constituting 700 us of each
forward pass). The idea is to implement the KVCollection construction as
a C symbol which we'll expose in the Python code and call into. We won't
eliminate all of this overhead, but we expect it to reduce.

Towards KERN-999

NN_ORIG_REV_ID: 76cbaad286dd0c46dc7f2e6cab2bafa9476d73b9