[Kernel][layout] Added initial CPU back-to-back matmul (D = A*B*C) test. (#47772)
The example uses a tiled layout that aims to maximize reuse of cache,
although blocking parameters are untuned (and would need to be for each
cache hierarchy).

The primary goal is to compute a tile of `A*B` at a time, storing it in
stack memory on the L3 cache, and using it to update all elements of the
corresponding rows of `D`.

I also use a strided access pattern within the microkernel.
In the CPU microkernel, we do an outer-prodcut between vector loads of
one matrix, and (broadcasted) scalars of the other.
That is, the microkernel calculates `C[M_r,W*N_r] +=
A[M_r,K_c]*B[K_c,W*N_r]`.

The microkernel is called in a loop that changes `N_r`, swapping out
tiles from `C` and `B`.

The matrix A uses `ceildiv(M_r * K_c * sizeof(float32),
number_of_sets*cacheline_size)` ways per set.
The matrix B uses `ceildiv(N_r * W * K_c * sizeof(float32),
number_of_sets*cacheline_size)` ways per set.

When the `ceildiv` rounds up, that means we're not using that many ways
from each set of the cache.

For my work laptop, I have:
12 total ways per set in the L1 cache
M_r = 6
N_r = 2
W = 8
number_of_sets = 64
cachline_size = 64

which yields a maximum value of K_c = 512, before we start evicting `A`
from the L1 cache, despite warning to keep it.

512 is fine/rather large.
For a Skylake-X machine, however:
8 total ways per set in the L1 cache
M_r = 9
N_r = 3
W = 16
number_of_sets = 64
cachline_size = 64

which yields a maximum of K_c = 113, which is on the small side as far
as blocking parameters go.

Thus, I took the approach here of accessing only 1 element per cacheline
of `A` at a time, but re-accessing them `cacheline_size /
sizeof[float32]()` times. This keeps the data in `A` away from the least
recently used position, and allows larger `K_c` values before we start
spilling `A` from the L1 cache (we do want to spill `B` from the L1
cache, as we aren't accessing `B` in the next call).

The re-accesses roughly let us divide the size of `B` by the
reaccess-frequency, to get `K_c = 560` as the largest possible value for
Skylake-X, up from 113 (note: `K_c` should now be a multiple of the
reaccess count, for ease of implementation in the micro-kernel).

This approach has the secondary benefit in back-to-back matrix multiply,
that we can store `AB = A*B` with contiguous vector stores while
producing that for repeated cache accesses in `AB * C`.

I'm not aware of anyone else using the same trick. I had been thinking
aout it for a while and implemented it in my polyhedral compiler's cost
model, but I haven't written codegen for it yet and thus hadn't been
able to test it. It also isn't particularly useful on my work laptop,
where a `K_c = 512` without using the trick is large enough.
Anyway, I thought it was an interesting enough idea that it may be worth
some consideration.

Mojo's layouts are useful for describing the memory layouts of `A`, `B`,
`C`, and `D` needed for this.

---------

Co-authored-by: abdul dakkak <adakkak@modular.com>
LAYOUT_ORIG_REV_ID: f52c98ac1347e80ba4b4b024ea44352c39f894d2