[Serve] Add README with license info (#48119)
We ship the Llama 3.1 8B model weights with the MAX Serve container
image. To comply with the licensing terms we need to add that info in
the README.md file and link the Llama license agreement.

Note: other sections of the MAX Serve README will be added later when we
get closer to the release and have its docs written.

Closes
[Internal link]

SERVE_ORIG_REV_ID: 737f7bffc04f791ff85e17476cb6f4fb26912795