[Serving] Expose task errors and die on error (#48141)
If a task completes, check to see if it completed due to unhandled
exception. If so, log the exception & stop the server.

I'm not positive that this error is unrecoverable, so please review for
that. The error should be exposed regardless.

SERVE_ORIG_REV_ID: 425b9c68d636da64c2d02b89700cb274819aff9e