[mojo-stdlib] Rename `memory.Reference` to `memory.Pointer`.
This renames the `Reference` struct to `Pointer`, to reflect this
type as being related to `UnsafePointer` (more rationale in the
changelog).

This patch is the minimal rename of the type itself without
updating the rest of the stdlib.  It includes an alias
to keep existing code working and easier to migrate.

This is the first step of MSTDL-948

GPU_ORIG_REV_ID: 91ed6a6480a2b07714c14bf408de39cdf0ed17d1