[Kernels] Converts 8xint4 (packed in 32-bits) to 8xbfloat16 ideal for Ampere (#48229)
Converts 8xint4 (packed in 32-bits) to 8xbfloat16 ideal for Ampere. The
output is interleaved so that the even values are the lower 4xint4
values and the odd values are the upper 4xint4 values. e.g. input `Int32
= 0x76543210` outputs `SIMD[DType.bfloat16, 8] = 0.0, 4.0, 1.0, 5.0,
2.0, 6.0, 3.0, 7.0`.

The lop3 instruction is currently emulated and will be added in a later
version.

Algorithm based on

https://github.com/PaddlePaddle/Paddle/blob/3862f8303d2723c03ffb42ce332d4c570906669f/paddle/phi/kernels/funcs/weight_only_gemv.cu#L795

---------

Co-authored-by: abdul dakkak <adakkak@modular.com>
MATH_ORIG_REV_ID: ea70a5685134464120d3c0749571dcdb380ceccb