[Benchmarking][Serving][Perf] Adds a lock on the fake "gpu" to prevent concurrent usage (#48960)
[Internal link] made the performance
fake's wait async, which released the CPU, which is good! However, it
allowed the fake to make multiple concurrent requests of the fake GPU,
which isn't realistic.

This adds a lock on the wait method to ensure that there isn't
concurrent usage of the fake GPU. There can be concurrent usage of the
fake GPU and the CPU; just not of the fake GPU and the fake GPU.

---------

Co-authored-by: Gaz Iqbal <gaz@modular.com>
SERVE_ORIG_REV_ID: ece6f1e056cac9978cb52eb7bf4b68d6af50fe7f