[KGEN][MOGGPreElab] Support Annotating the Number of DPS Operand Outputs (#49092)
This PR adds a new operand to mojo `@register` which specifies the number of DPS operands and forward this information to the kgen generator.

The is useful for supporting multiple DPS results or zero DPS results, which is useful for opaque type support.

EXTENSIBILITY_ORIG_REV_ID: 069906da220d3823f918cac8a211dc52bd9df588