[Tracing] Expose the trace marker to the SDK (#49307)
This exposes trace markers to the SDK. The markers are different from
spans because they signal the state at a point in time vs a span which
signals the state at a region of time. For example, one might register
the marker as a backend to the logging and/or measure to populate the
profiler.

---------

Co-authored-by: Brendan Duke <bduke@modular.com>
RUNTIME_ORIG_REV_ID: 8dc33b886ef41ad748ad8cf131474681725db73b