[Kernels][GPU] Rename compileFunction to loadFunction in device context (#50348)
While technically correct that you are compiling the function, it is
more correct to say that you are loading the function with compilation
being a byproduct for NVIDIA GPUs. But, for AMD GPUs this a purly
loading the function.

This just does the rename to better reflect what the function is doing.

GPU_ORIG_REV_ID: 9eabc14ddfb47f70bad5eb2c2dbe121e510d2146