[Benchmarking] Add option to record and dump batch info from perf fake (#50529)
Every time the perf fake is called, it stores the shapes in the batch.
At shutdown, it dumps those to a json file. This only occurs when the
filename is set at fake init time because the record action is in a
performance critical region of the code.

SERVE_ORIG_REV_ID: d097d2ae569596452ef4a477544d378870ac146a