[Serve] Add new Engine process worker (#50608)
This PR is a part of the effort to move batching logic to the Engine
process. It adds a new Engine process model worker.
It combines context encoding (dynamic_batching_worker) and token
generation (continuous_batching_worker) into a single loop and moves it
inside the Engine process. This is our decision and extensibility point
to decide what we what to schedule on GPU and for how long (num steps).
The goal is to have the engine loop as tight as possible without
unnecessary idling and waiting for request/response processing or any
other Python work unrelated to GPU.

This improves performance quite a bit:
QPS of the perf fake model with busy wait increases by 22%
QPS of the Llama3 model increases by 7% (~12% if running for longer
time)

More info can be found here
[Internal link]

Note: This PR does not enable the new model yet. It's just code for now.
We will enable it in the subsequent PRs.

Related: [Internal link]
SERVE_ORIG_REV_ID: 033dcb91036870da6903b418fc9526072ce5a94f