[Extensibility] Fix reduction kernels for bf16 types (#50769)
The extensibility kernels for reduction operations min, max, sum, and
product are written in such a way that the input/output functions are
prematurely specialized on the DType of the data being processed.

On the GPU, an accumulator type of f32 is used for half precision types
(f16 & bf16). When the accumulator type does not match the input data
type, elaboration fails for the kernels.

Closes: KERN-1195
ALGORITHM_ORIG_REV_ID: 1930b8059f1f6806ad5cb6796aef613eb662271a