[KGEN][Kernels] Compute the function hash based on the function body (#50953)
Before we were hashing just the name of the function. This worked for
a while until it didn't. The issues is described in GRA-1271, but we
can have a collision where two functions have the same name, are
parametrized in the same way, but have different bodies.

So, make the key of the function cache based on the body of the function
being cached rather than using the function name. This is done by
introducing a new string.hash compile op that gives the blake3 hash
of a string literal.

We cannot use the string literal hash function, since that returns a
UInt and will have massive hash collisions.

Closes GRA-1271.

GPU_ORIG_REV_ID: 83614567a3b7e2a2a68d0e97e407886756c93c2a