[mojo-lang] Substitute aliases in function bodies eagerly. (#51002)
This changes the Mojo parser to substitute aliases in function bodies
eagerly, like it does for global aliases and structs. This fixes a whole
class of issues that forced users to use rebind and other workarounds
due to the compiler not understanding that types are identical due to
local alias mismatches.  This original requirement driving this dates to
ancient times which is no longer relevant to today.

This fixes MOCO-1440 and probably others.

GPU_TEST_ORIG_REV_ID: 0211cd8c7fa0b2daf6e512d26fcd07c74f1095da