[Kernels] bug fix for the MHA naive GPU kernel (#51497)
The order of the indices used for the mask functor were reversed in the
MHA naive path which right now is only hit on AMD GPUs. Forcing the
nvidia path to use the naive kernel also hit the same problem. The row
offset also needs to be biased by the cache length as is done in other
uses of this mask functor.

Also fixed a test to build the correct mask for the reference/fallback
code path. Each batch has a different cache length and that needs to be
baked into the mask tensor. This is not typically seen because of the
kernel_mask doing yet another layer of masking on top of this. I was
hitting this while attempting to debug the above.

NN_ORIG_REV_ID: 973b1eb727f48d645ba4eef41b814cb2983cd31c