[Pipelines] Update Safetensor to support loading weight as uint8 bytes (#51663)
Stacked PRs:
 * __->__#51663

--- --- ---

### [Pipelines] Update Safetensor to support loading weight as uint8
bytes

This is needed for the quantization work as we sometimes need to load
the weights as just bytes and the graph compiler and the graph SDK
currently lack the ability to do a bitcast. Even if they did have the
ability to do a bitcast, this might be the better option, in terms of
compilation time.

MAX_GRAPH_API_ORIG_REV_ID: a8c0801c46a4e65b3bf3abf3eb0fd2aca1ffd949